requirements.txt
cffi==1.14.2
click==7.1.2
cryptography==3.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
PyMySQL==0.10.0
PyYAML==5.3.1
six==1.15.0
Werkzeug==1.0.1
